ACCUMULO-1783 Clean up the "typed" variant from before and supersede the original AccumuloStorage.
Going off of what HBaseStorage provides, the typical usecase is
treating each tuple as a "row" and inserting the multiple columns from a
bag or likewise as columns in that row. Trying to move towards this, I
renamed the AccumuloStorage into AccumuloKVStorage to make it obvious
that this storage engine is purely providing a Key/Value storage
mechanism and nothing more.