ACCUMULO-1783 Reworking the storage side.
Took an approach like HBase's for the "regular" AccumuloStorage class.
Normal tuples are treated as a row, with the first entry being the
rowkey and subsequent entries as column values. Maps are expanded as
column:value pairs, any scalars, bags or tuples require a column in the
AccumuloStorage constructor. Lots of nice unit tests for the
functionality